popups:
- internal implementation details & all service declarations moved to m_popup_int.h;
- all service calls removed and replaced with function calls;
- direct access to popup serttings replaced with Popup_Enable / Popup_Enabled;

+++ plugins/Exchange/src/emails.cpp
@@ -312,7 +312,7 @@ int ShowPopupMessage(wchar_t *title, wchar_t *message, int cUnreadEmails)
 	wcsncpy_s(popup.lpwzText, MAX_SECONDLINE, message, _TRUNCATE);
 	popup.PluginWindowProc = DlgProcPopup;
 	popup.PluginData = (int *) cUnreadEmails;
-	return PUAddPopupW(&popup);
+	return (int)PUAddPopupW(&popup);
 }
 
 int ShowMessageBoxMessage(wchar_t *title, wchar_t *message, int cUnreadEmails)
